18_Capstone/gradio_deploy/requirements.txt
gradio==5.29.0 
# pip install --upgrade gradio
# pip show gradio - to check version on local 
torch 
torchvision 
boto3 
python-dotenv 
pillow 
psutil